requirements.txt
telethon
python-decouple
asyncio
